pyproject.toml
[project]
name = "claude-usage"
version = "0.0.14"
description = "macOS menu bar app for Anthropic API usage and cost tracking"
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Anders Latif" },
]
keywords = ["anthropic", "claude", "menu-bar", "macos", "usage", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]

dependencies = [
    "rumps>=0.4.0",
    "requests>=2.32.3",
    "keyring>=25.6.0",
    "python-dotenv>=1.0.1",
    "pyobjc-framework-quartz>=12.1",
    "pyobjc-framework-coretext>=12.1",
]

[project.scripts]
claude-usage = "src.cli:main"

[project.urls]
Repository = "https://github.com/anderslatif/claude-usage"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
